3기_도서관대출_김민석/requirements.txt
aiohttp==3.7.4.post0
alembic==1.7.4
aniso8601==9.0.1
APScheduler==3.8.1
async-timeout==3.0.1
attrs==21.2.0
Automat==0.8.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blinker==1.4
blueprint==3.4.2
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.3
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
discord==1.7.3
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.4.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.1.5
MarkupSafe==2.0.1
more-itertools==4.2.0
multidict==5.2.0
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.18.0
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sos==4.1
SQLAlchemy==1.3.24
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.3
wadllib==1.3.3
Werkzeug==2.0.2
WTForms==3.0.0
yarl==1.7.0
zipp==3.6.0
zope.interface==4.7.1
